Add mark_read method and return number of unread messages when calling list_chats

+++ src/chats/src/updates/send_message.rs
@@ -2,7 +2,7 @@ use ic_cdk::{api, storage};
 use crate::domain::chat::ChatId;
 use crate::domain::chat_list::ChatList;
 
-pub fn update(chat_id: ChatId, text: String) -> Option<u64> {
+pub fn update(chat_id: ChatId, text: String) -> Option<u32> {
     let chat_list: &mut ChatList = storage::get_mut();
     let me = ic_cdk::caller();
